v1.1.20: restructure code, new method: getAllAccountBalance, fix Date Type
