+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@brickglobal/brick-sdk",
-  "version": "1.1.19",
+  "version": "1.1.20",
   "main": "dist/index.js",
   "files": [
     "dist"
