+++ src/errorCode.ts
@@ -1,0 +1,7 @@
+
+export const BMErrorCode = {
+    PARAM_INVALID: `PARAM_INVALID`,
+    PARAM_MISSING: `PARAM_MISSING`,
+    API_KEY_MISSING: `API_KEY_MISSING`,
+    PROVIDER_MISSING: `PROVIDER_MISSING`
+}
