+++ src/index.ts
@@ -1,3 +1,4 @@
-import { BrickSDK, Transaction, SubAccount, MainAccount, RecheckResult } from "./brickSDK";
+import { BrickSDK } from "./brickSDK";
+import { MainAccount, RecheckResponse, SubAccount, Transaction } from "./type";
 
-export { BrickSDK, Transaction, SubAccount, MainAccount, RecheckResult }
+export { BrickSDK, Transaction, SubAccount, MainAccount, RecheckResponse }
