+++ src/methods/method.creditSub.ts
@@ -1,4 +1,4 @@
-import { BMApolloMethodName } from "../brickSDK"
+import { BMApolloMethodName } from "."
 
 export const creditSub = (params: any): { name: string, query: string } => {
     return {
