+++ src/methods/method.debitSub.ts
@@ -1,4 +1,4 @@
-import { BMApolloMethodName } from "../brickSDK"
+import { BMApolloMethodName } from "."
 
 export const debitSub = (params: any): { name: string, query: string } => {
     return {
