+++ src/methods/method.getAllAccountBalance.ts
@@ -1,0 +1,16 @@
+import { BMApolloMethodName } from "."
+
+
+export const getAllAccountBalance = (params: any): { name: string, query: string } => {
+    return {
+        name: BMApolloMethodName.getAllAccountBalance,
+        query: `
+            query {
+                ${BMApolloMethodName.getAllAccountBalance} {
+                eur
+                usdt_trc20
+                }
+            }
+          `
+    }
+}
