+++ src/methods/method.getMainAccInfo.ts
@@ -1,4 +1,4 @@
-import { BMApolloMethodName } from "../brickSDK"
+import { BMApolloMethodName } from "."
 
 export const getMainAccInfo = (params: any): { name: string, query: string } => {
     return {
