+++ src/methods/method.recheckTx.ts
@@ -1,4 +1,4 @@
-import { BMApolloMethodName } from "../brickSDK"
+import { BMApolloMethodName } from "."
 
 export const recheckTx = (params: any): { name: string, query: string } => {
     return {
