+++ src/methods/method.subAccountInfo.ts
@@ -1,4 +1,4 @@
-import { BMApolloMethodName } from "../brickSDK"
+import { BMApolloMethodName } from "."
 
 export const subAccountInfo = (params: any): { name: string, query: string } => {
     return {
