+++ src/type/MethodResponses.ts
@@ -1,0 +1,10 @@
+export type RecheckResponse = {
+    txid: string
+    status: string
+}
+
+export type AllAccountBalanceResponse = {
+    eur: number
+    usdt_trc20: number
+}
+
