+++ src/type/index.ts
@@ -1,0 +1,12 @@
+import { MainAccount } from "./MainAccount"
+import { SubAccount } from "./SubAccount"
+import { Transaction } from "./Transaction"
+import { RecheckResponse, AllAccountBalanceResponse } from "./MethodResponses"
+
+export {
+    MainAccount,
+    SubAccount,
+    Transaction,
+    RecheckResponse,
+    AllAccountBalanceResponse
+}
